fix errors, implement LFA feedback, make heroku return interface [Delivers #165239420]
