+++ app/views/static/js/removemember.js
@@ -1,60 +1,49 @@
-let email = document.getElementById("user_email");
+let user_id = document.getElementById("user_id");
 let group_id = document.getElementById("group_id");
 let token = localStorage.getItem('token');
 console.log(token);
 let url = "";
 
-function email_validation(){
-    let email_error = document.getElementById("email_error");
-    if (/\S+@\S+\.\S+/.test(email.value)){
-        email_error.style.display = "none";
-        email.setCustomValidity("")
-    }
-    else {
-        email_error.style.display = "block";
-        email_error.style.color = "red";
-        email_error.innerHTML = "Enter correct email";
-        email.setCustomValidity("Wrong Email Format.")
+function logout() {
+    localStorage.removeItem('token');
+    window.location.replace("./index.html");
+}
+window.onload = function not_signedin(){
+    if(token === null){
+        window.location.replace("./index.html");
     }
 }
-email.onkeyup = email_validation;
-email.onchange = email_validation;
 
 group_id.onkeyup = function group_id_validation() {
-    let email_error = document.getElementById("gid_error");
-    if (/^[1-9]{1,}$/.test(group_id.value)) {
-        email_error.style.display = "none";
-        group_id.setCustomValidity("");
-        url = "https://epicmail-sentongo-v2.herokuapp.com/api/v2/groups/"+group_id.value+"/users";
+    let uid_error = document.getElementById("uid_error");
+    if (/^[0-9]{1,}$/.test(group_id.value)) {
+        uid_error.style.display = "none";
+        user_id.setCustomValidity("");
+        url = "https://epicmail-sentongo-v2.herokuapp.com/api/v2/groups/"+group_id.value+"/users/"+user_id.value;
         console.log(url)
     } else {
-        email_error.style.display = "block";
-        email_error.style.color = "red";
-        email_error.innerHTML = "Group ID must be an integer";
-        group_id.setCustomValidity("Invalid Group ID.");
+        uid_error.style.display = "block";
+        uid_error.style.color = "red";
+        uid_error.innerHTML = "User ID must be an integer";
+        user_id.setCustomValidity("Invalid User ID.");
     }
 }
 
-function add_member() {
-    let user = {
-        user_email: email.value
-    };
-
+function remove_member() {
     fetch(url, {
-        method: "POST",
+        method: "DELETE",
         headers: {
             "content-type": "application/json",
             "Authorization": `Bearer ${token}`
         },
-        body: JSON.stringify(user)
     })
         .then((response) => response.json())
         .then((data) => {
             if (data.status === 200) {
                 document.getElementById("page_response").style.display = "block";
                 document.getElementById("page_response").style.color = "green";
-                document.getElementById("page_response").innerHTML = "User Added";
-                window.location.replace("./inbox.html");
+                document.getElementById("page_response").innerHTML = "User Removed";
+                window.location.replace("./mygroups.html");
             }
             else if (data.status === 400) {
                 document.getElementById("page_response").style.display = "block";
@@ -71,5 +60,6 @@ function add_member() {
             else if (data.msg === "Token has expired") {
                 window.location.replace("./index.html");
             }
+
         })
 }
