+++ app/views/templates/drafts.html
@@ -3,27 +3,26 @@
 <head>
     <meta charset="utf-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>EpicMail || inbox</title>
+    <title>EpicMail || drafts</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" media="screen" href="static/main.css" />
-    <script src="static/main.js"></script>
+    <link rel="stylesheet" type="text/css" media="screen" href="../static/main.css" />
+    <script src="../static/main.js"></script>
 </head>
 <body>
     <header>
             <div class="container">
                 <div class="logo">
                     
-                    <a href="inbox.html">
-                            <img src="static/img/epic.png" alt="EpicMail Logo" height="70" >
+                    <a href="drafts.html">
+                            <img src="../static/img/epic.png" alt="EpicMail Logo" height="70" >
                         </a>
                         <h1>EpicMail</h1> 
                 </div>
-                
             </div>
         </header>
 
     <section id="page" >
-        <h3 id="heading">INBOX</h3>
+        <h3 id="heading">DRAFTS</h3>
         <nav id="sidenav">
             <ul>
                 <li ><a href="inbox.html">Inbox</a></li>
@@ -36,12 +35,12 @@
                 <li ><a href="mygroups.html">My Groups</a></li>
                 <li ><a href="editname.html">Rename Group</a></li>
                 <li ><a href="delete_member.html">Remove Member</a></li>
-                <li ><a href="index.html">Sign Out</a></li>
+                <li ><a href="#" onclick="logout()">Sign Out</a></li>
             </ul>
         </nav>
     <div id="mailist">
-  
-    </div>
+        
+    </div>   
     </section>
     <footer>
         <div class="footer-info">
@@ -51,5 +50,5 @@
     </footer>
     
 </body>
-<script src="static/js/inbox.js"></script>
+<script src="../static/js/drafts.js"></script>
 </html>
