+++ app/views/templates/editname.html
@@ -3,27 +3,26 @@
 <head>
     <meta charset="utf-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>EpicMail || my groups</title>
+    <title>EpicMail || Rename Group</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" media="screen" href="static/main.css" />
-    <script src="static/main.js"></script>
+    <link rel="stylesheet" type="text/css" media="screen" href="../static/main.css" />
+    <script src="../static/main.js"></script>
 </head>
 <body>
     <header>
             <div class="container">
                 <div class="logo">
                     
-                    <a href="mygroups.html">
-                            <img src="static/img/epic.png" alt="EpicMail Logo" height="70" >
+                    <a href="editname.html">
+                            <img src="../static/img/epic.png" alt="EpicMail Logo" height="70" >
                         </a>
                         <h1>EpicMail</h1> 
                 </div>
-                
             </div>
         </header>
 
     <section id="page" >
-        <h3 id="heading">MY GROUPS</h3>
+        <h3 id="heading">EDIT GROUP NAME</h3>
         <nav id="sidenav">
             <ul>
                 <li ><a href="inbox.html">Inbox</a></li>
@@ -36,18 +35,25 @@
                 <li ><a href="mygroups.html">My Groups</a></li>
                 <li ><a href="editname.html">Rename Group</a></li>
                 <li ><a href="delete_member.html">Remove Member</a></li>
-                <li ><a href="index.html">Sign Out</a></li>
+                <li ><a href="#" onclick="logout()">Sign Out</a></li>
+            </ul>
         </nav>
     <div id="mailist">
-  
-    </div>
+        <div class="mailinputs">
+            <form action="#" method="GET" onsubmit="rename(); return false">
+                <br><br>
+                <input type="text" placeholder="New Name ...." required id="new_name"><span id="name_error"></span>
+                <input type="text" placeholder="Group ID ...." required id="group_id"><span id="gid_error"></span><br><span id="page_response"></span>
+                <button type="submit" class="button1" id="add">RENAME</button>           <button class="button1" type="reset">Discard</button>
+            </form>
+        </div>
     </section>
     <footer>
         <div class="footer-info">
-            <b><p id="footext">All Rights reserved. T&Cs apply. </p></b>
+            <b><p id="footext">All Rights reserved. T&Cs apply. </p>
         </div>
     </footer>
     
 </body>
-<script src="static/js/mygroups.js"></script>
+<script src="../static/js/editname.js"></script>
 </html>
