+++ app/views/templates/groupmail.html
@@ -5,17 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>EpicMail || create group</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" media="screen" href="static/main.css" />
-    <script src="static/main.js"></script>
-</head><!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>EpicMail || compose</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" media="screen" href="static/main.css" />
-    <script src="static/main.js"></script>
+    <link rel="stylesheet" type="text/css" media="screen" href="../static/main.css" />
+    <script src="../static/main.js"></script>
 </head>
 <body>
     <header>
@@ -23,7 +14,7 @@
                 <div class="logo">
                     
                     <a href="groupchat.html">
-                            <img src="static/img/epic.png" alt="EpicMail Logo" height="70" >
+                            <img src="../static/img/epic.png" alt="EpicMail Logo" height="70" >
                         </a>
                         <h1>EpicMail</h1> 
                 </div>
@@ -38,13 +29,13 @@
                 <li ><a href="new.html">Compose</a></li>
                 <li ><a href="sent.html">Sent</a></li>
                 <li ><a href="drafts.html">Drafts</a></li>
-                <li ><a href="group.html">Create Group</a></li>
+                <li ><a href="groupmail.html">Create Group</a></li>
                 <li ><a href="groupmail.html">Group Chat</a></li>
                 <li ><a href="addmembers.html">Add Member</a></li>
                 <li ><a href="mygroups.html">My Groups</a></li>
                 <li ><a href="editname.html">Rename Group</a></li>
                 <li ><a href="delete_member.html">Remove Member</a></li>
-                <li ><a href="index.html">Sign Out</a></li>
+                <li ><a href="#" onclick="logout()">Sign Out</a></li>
             </ul>
         </nav>
     <div id="mailist">
@@ -65,5 +56,5 @@
     </footer>
     
 </body>
-<script src="static/js/send_group_mail.js"></script>
+<script src="../static/js/send_group_mail.js"></script>
 </html>
