+++ app/views/templates/new.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>EpicMail || compose</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" media="screen" href="static/main.css" />
-    <script src="static/main.js"></script>
+    <link rel="stylesheet" type="text/css" media="screen" href="../static/main.css" />
+    <script src="../static/main.js"></script>
 </head>
 <body>
     <header>
@@ -14,7 +14,7 @@
                 <div class="logo">
                     
                     <a href="new.html">
-                            <img src="static/img/epic.png" alt="EpicMail Logo" height="70" >
+                            <img src="../static/img/epic.png" alt="EpicMail Logo" height="70" >
                         </a>
                         <h1>EpicMail</h1> 
                 </div>
@@ -35,7 +35,7 @@
                 <li ><a href="mygroups.html">My Groups</a></li>
                 <li ><a href="editname.html">Rename Group</a></li>
                 <li ><a href="delete_member.html">Remove Member</a></li>
-                <li ><a href="index.html">Sign Out</a></li>
+                <li ><a href="#" onclick="logout()">Sign Out</a></li>
             </ul>
         </nav>
     <div id="mailist">
@@ -56,5 +56,5 @@
     </footer>
     
 </body>
-<script src="static/js/sendmail.js"></script>
+<script src="../static/js/sendmail.js"></script>
 </html>
