+++ app/views/templates/signup.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>EpicMail || Sign Up</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" media="screen" href="static/main.css" />
-    <script src="static/main.js"></script>
+    <link rel="stylesheet" type="text/css" media="screen" href="../static/main.css" />
+    <script src="../static/main.js"></script>
 </head>
 <body>
     <header>
@@ -14,7 +14,7 @@
                 <div class="logo">
                     
                     <a href="signup.html">
-                            <img src="static/img/epic.png" alt="EpicMail Logo" height="70" >
+                            <img src="../static/img/epic.png" alt="EpicMail Logo" height="70" >
                         </a>
                         <h1>EpicMail</h1> 
                 </div>
@@ -58,5 +58,5 @@
     </footer>
     
 </body>
-<script src="static/js/signup.js"></script>
+<script src="../static/js/signup.js"></script>
 </html>
