+++ index.html
@@ -1,0 +1,16 @@
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>WELCOME</title>
+    <script>
+        function go_to_home(){
+            window.location.replace("./app/views/templates/index.html");
+        }
+    </script>
+</head>
+<body onload="go_to_home()">
+    
+</body>
+</html>
